draging and droping is working fine
